fix(accessibility): respect prefers-reduced-motion in dashboard header animations
Extract reduced motion detection into usePrefersReducedMotion hook and apply it to DashboardHeader to disable dotFlash/alertPulse animations when users prefer reduced motion. Aligns behavior with sidebar.
